+++ play-frontend-hmrc-play-30/src/main/scala/uk/gov/hmrc/hmrcfrontend/config/TrackingConsentConfig.scala
@@ -26,8 +26,6 @@ class TrackingConsentConfig @Inject() (config: Configuration) {
     platformHost.map(_ => "").orElse(config.getOptional[String]("tracking-consent-frontend.host"))
   val trackingConsentPath: Option[String] =
     config.getOptional[String]("tracking-consent-frontend.path")
-  val optimizelyGtmPath: Option[String]   =
-    config.getOptional[String]("tracking-consent-frontend.optimizely-gtm-path")
   val gtmContainer: Option[String]        = config.getOptional[String]("tracking-consent-frontend.gtm.container")
 
   def trackingUrl(): Option[String] =
@@ -42,11 +40,4 @@ class TrackingConsentConfig @Inject() (config: Configuration) {
       baseUrl   <- config.getOptional[String]("optimizely.url")
       projectId <- config.getOptional[String]("optimizely.projectId")
     } yield s"$baseUrl$projectId.js"
-
-  def optimizelyGtmUrl(): Option[String] =
-    for {
-      host <- trackingConsentHost
-      path <- optimizelyGtmPath
-      _    <- gtmContainer
-    } yield s"$host$path"
 }